requirements.txt
fastapi==0.135.3
uvicorn==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
PyMuPDF==1.27.2.2
python-docx==1.2.0
python-multipart==0.0.22
groq==1.1.2
httpx==0.28.1
chromadb==0.5.23
